requirements.txt
torch>=1.7.1
torchvision>=0.8.2
numpy
opencv-python-headless
Pillow
scipy
pytest
pytest-mock
scikit-learn
pre-commit
